+++ src/main/java/mhel/itu/bachelor/shortestpathmap/algorithm/IVertex.java
@@ -2,6 +2,6 @@ package mhel.itu.bachelor.shortestpathmap.algorithm;
 
 public interface IVertex {
     int I();
-    float X();
-    float Y();
+    double X();
+    double Y();
 }
